backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
firebase-admin==6.4.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-dotenv==1.0.1
gunicorn==21.2.0
google-auth>=2.0.0
requests>=2.31.0
python-jose[cryptography]>=3.3.0
